+++ packages/components-react/navigation-bar-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/navigation-bar-css';
+
+export * from './NavBar';