+++ packages/components-react/unordered-list-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/unordered-list-css';
+
+export * from './UnorderedList';