+++ packages/storybook/src/templates/design-tokens/LineHeightSample/index.tsx
@@ -12,7 +12,7 @@ export const LineHeightSample = ({
 }: PropsWithChildren<
   {
     fontFamily: string;
-    fontWeight: string;
+    fontWeight: number;
     fontSize: string;
     lineHeight: string;
   } & HTMLAttributes<HTMLDivElement>