+++ packages/components-react/card-react/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @rijkshuisstijl-community/card-react
 
+## 1.1.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+- Updated dependencies [e41c59f]
+  - @rijkshuisstijl-community/link-react@1.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- de35f24: Update README files
+- c9e30de: Added an option to import the components separate from the css.
+
+### Patch Changes
+
+- c414601: Ensure package is published with provenance
+- 02786d9: security patch
+- Updated dependencies [c414601]
+- Updated dependencies [02786d9]
+- Updated dependencies [de35f24]
+- Updated dependencies [c9e30de]
+  - @rijkshuisstijl-community/link-react@1.1.0
+  - @rijkshuisstijl-community/card-css@1.1.0
+
 ## 1.0.0
 
 ### Major Changes