+++ packages/components-react/checkbox-group-react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @rijkshuisstijl-community/checkbox-group-react
 
+## 1.1.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+
+## 1.1.0
+
+### Minor Changes
+
+- c9e30de: Added an option to import the components separate from the css.
+
+### Patch Changes
+
+- c414601: Ensure package is published with provenance
+- 02786d9: security patch
+- Updated dependencies [c414601]
+  - @rijkshuisstijl-community/checkbox-group-css@1.0.1
+
 ## 1.0.0
 
 ### Major Changes